requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=2.0.0=pypi_0
alabaster=0.7.13=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.4.0=pyhd8ed1ab_0
babel=2.13.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
bleach=6.1.0=pypi_0
boto3=1.33.5=pypi_0
botocore=1.33.5=pypi_0
ca-certificates=2023.7.22=h8857fd0_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.7=py38hcec6c5f_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
docutils=0.16=pypi_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=1.4=pypi_0
filelock=3.12.4=pypi_0
fonttools=4.42.1=pypi_0
fsspec=2023.10.0=pypi_0
futures=3.0.5=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
google-auth=2.23.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.58.0=pypi_0
gtn=0.0.0=pypi_0
idna=3.4=pypi_0
imagesize=1.4.1=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.1.0=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.25.2=pyh1050b4e_0
ipython=8.12.0=pyhd1c38e8_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=5.3.1=py38h50d1736_0
kaggle=1.5.16=pypi_0
kiwisolver=1.4.5=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
libsodium=1.0.18=hbcb3906_1
lightning-utilities=0.10.0=pypi_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.76=pypi_0
openssl=3.1.3=h8a1eda9_0
packaging=23.1=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.3.0=pypi_0
pip=23.2.1=py38hecd8cb5_0
platformdirs=3.10.0=pyhd8ed1ab_0
promise=2.3=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.24.3=pypi_0
psutil=5.9.5=py38hcafd530_1
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
python=3.8.18=h5ee71fb_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-slugify=8.0.1=pypi_0
python_abi=3.8=2_cp38
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=24.0.1=py38h0b711fd_0
readline=8.2=hca72f7f_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
s3=3.0.0=pypi_0
s3transfer=0.8.2=pypi_0
scipy=1.10.1=pypi_0
sentry-sdk=1.38.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.0.0=py38hecd8cb5_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
snowballstemmer=2.2.0=pypi_0
sphinx=5.3.0=pypi_0
sphinx-rtd-theme=0.5.2=pypi_0
sphinxcontrib-applehelp=1.0.4=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.1=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
text-unidecode=1.3=pypi_0
tk=8.6.12=h5d9f67b_0
torch=2.1.1=pypi_0
torchaudio=2.1.1=pypi_0
torchmetrics=1.2.1=pypi_0
torchvision=0.16.1=pypi_0
tornado=6.3.3=py38hcafd530_1
tqdm=4.64.1=pypi_0
traitlets=5.10.0=pyhd8ed1ab_0
typing-extensions=4.7.1=pypi_0
typing_extensions=4.8.0=pyha770c72_0
urllib3=1.26.18=pypi_0
wandb=0.13.5=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py38hecd8cb5_0
xmltodict=0.13.0=pypi_0
xz=5.4.2=h6c40b1e_0
zeromq=4.3.4=he49afe7_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h4dc903c_0
